Move formatDate into stripHelpers. Do date formatting in middleware
