+++ src/routes/supplierRoutes.ts
@@ -8,47 +8,10 @@ import axios from "axios";
 import { createAbacMiddleware } from "../middleware/ABACMiddleware";
 import { shareRequestDetailMiddleware } from "../middleware/apiMiddleware";
 import { replace } from "../helperFunctions/checkhelper";
-import { capitalise } from "../helperFunctions/stringHelpers";
+import { capitalise, formatDate } from "../helperFunctions/stringHelpers";
+import { checkAnswer } from "../helperFunctions/formHelper";
 const router = express.Router();
 const URL = `${process.env.API_ENDPOINT}/manage-shares`;
-
-const formatDate = (dateString: string): string => {
-  if (dateString === "UNREQUESTED") {
-    return "Unrequested";
-  }
-  const options: Intl.DateTimeFormatOptions = {
-    year: "numeric",
-    month: "long",
-    day: "numeric",
-  };
-  return new Date(dateString).toLocaleDateString("en-GB", options);
-};
-
-function formatDateObject(dateObj: {
-  day: string;
-  month: string;
-  year: string;
-}): string {
-  const monthNames = [
-    "January",
-    "February",
-    "March",
-    "April",
-    "May",
-    "June",
-    "July",
-    "August",
-    "September",
-    "October",
-    "November",
-    "December",
-  ];
-
-  const monthIndex = parseInt(dateObj.month, 10) - 1; // Convert month string to number and subtract 1 for zero-based index
-  const monthName = monthNames[monthIndex];
-
-  return `${dateObj.day} ${monthName} ${dateObj.year}`;
-}
 
 router.get("/", async (req: Request, res: Response) => {
   const backLink = req.headers.referer || "/";
@@ -127,8 +90,7 @@ router.get(
       { text: r.assetTitle },
       { text: r.assetPublisher.title },
       {
-        text:
-          r.neededBy === "UNREQUESTED" ? "Unrequested" : formatDate(r.neededBy),
+        text: formatDate(r.neededBy),
       },
       {
         html: `<span class="govuk-tag ${getCreatedRequestStatusClass(
@@ -208,39 +170,37 @@ router.get(
 
 router.get(
   "/created-requests/:requestId",
+  shareRequestDetailMiddleware,
   async (req: Request, res: Response) => {
     const backLink = req.headers.referer || "/manage-shares/created-requests/";
-    const requestId = req.params.requestId;
 
     try {
-      const requestDetail = await axios.get(
-        `${URL}/received-requests/${requestId}`,
-        {
-          headers: { Authorization: `Bearer ${req.cookies.jwtToken}` },
-        },
-      );
-
-      if (!requestDetail.data) {
-        return res.status(404).send("Request not found");
-      }
-
-      // Format the received date
-      requestDetail.data.received = formatDate(requestDetail.data.received);
-      requestDetail.data.neededBy = formatDate(requestDetail.data.neededBy);
-      requestDetail.data.decisionDate = formatDate(
-        requestDetail.data.decisionDate,
-      );
-
-      // Format the neededBy date
-      const dateObj = requestDetail.data.sharedata.steps.date.value;
-      requestDetail.data.sharedata.steps.date.formattedValue =
-        formatDateObject(dateObj);
-
-      req.session.acquirerForms = requestDetail.data;
+      const requestDetail = req.shareRequest!;
 
       res.render("../views/acquirer/created-request-outcome.njk", {
         backLink,
-        request: requestDetail.data,
+        request: requestDetail,
+      });
+    } catch (error) {
+      console.error(error);
+      res.status(500).send("Internal Server Error");
+    }
+  },
+);
+
+router.get(
+  "/created-requests/:requestId/view-answers",
+  shareRequestDetailMiddleware,
+  async (req: Request, res: Response) => {
+    const backLink = req.headers.referer || `/manage-shares/created-requests/`;
+
+    try {
+      const requestDetail = req.shareRequest!;
+
+      res.render("../views/acquirer/created-request-view-full-request.njk", {
+        request: requestDetail,
+        sharedata: checkAnswer(requestDetail.sharedata),
+        backLink,
       });
     } catch (error) {
       console.error(error);
@@ -321,9 +281,7 @@ router.get(
       { text: r.requestingOrg },
       { text: r.assetTitle },
       { text: formatDate(r.received) },
-      {
-        text: r.decisionDate ? formatDate(r.decisionDate) : "",
-      },
+      { text: formatDate(r.decisionDate) },
       {
         html: `${capitalise(r.status)}`,
       },
@@ -351,9 +309,6 @@ router.get(
 
     try {
       const requestDetail = req.shareRequest!;
-      // Format the dates
-      requestDetail.received = formatDate(requestDetail.received);
-      requestDetail.neededBy = formatDate(requestDetail.neededBy);
 
       res.render("../views/supplier/review-summary.njk", {
         backLink,
@@ -383,7 +338,6 @@ router.get(
   shareRequestDetailMiddleware,
   (req: Request, res: Response) => {
     const requestDetail = req.shareRequest!;
-    requestDetail.neededBy = formatDate(requestDetail.neededBy);
 
     res.render("../views/supplier/review-request.njk", {
       request: requestDetail,
@@ -627,11 +581,6 @@ router.get(
 
     try {
       const requestDetail = req.shareRequest!;
-      // Format the received date
-      requestDetail.received = formatDate(requestDetail.received);
-      requestDetail.neededBy = formatDate(requestDetail.neededBy);
-      requestDetail.decisionDate = formatDate(requestDetail.decisionDate || "");
-      requestDetail.neededBy = formatDate(requestDetail.neededBy);
 
       res.render("../views/supplier/received-request-outcome.njk", {
         backLink,
@@ -656,9 +605,6 @@ router.get(
 
     try {
       const requestDetail = req.shareRequest!;
-      // Format the dates
-      requestDetail.received = formatDate(requestDetail.received);
-      requestDetail.neededBy = formatDate(requestDetail.neededBy);
 
       res.render("../views/supplier/received-request-view-full-request.njk", {
         request: requestDetail,
